move `devide_id` & `group_id` into their own files and add unit tests
